Conflicts:
	.gitignore

+++ app/controllers/drafts_controller.rb
@@ -1,20 +1,33 @@
 class DraftsController < ApplicationController
   before_action :authenticate_researcher!
 
-  EMPTY_SCENARIO = Scenario.new({
-    scenario_steps: [ScenarioStep.new()]
-  })
-
   def index
     @scenarios = Scenario.drafts(current_researcher.id)
   end
 
-  def new(scenario = EMPTY_SCENARIO)
+  def new
+    p 'calling new'
+    new_or_edit(Scenario.new({
+      scenario_steps: [ScenarioStep.new()],
+      status: :drafts
+    }))
+    render :new
+  end
+
+  def edit
+    p 'calling edit'
+    new_or_edit(Scenario.find_by_hashid(params[:id]))
+    render :new
+    # TODO: modify the new page to indicate it is editing
+  end
+
+  def new_or_edit(scenario)
+    p scenario
     @scenario = scenario
     @product_templates = Template.product_templates
     @marketing_templates = Template.marketing_templates
     if @template = Template.find_by(value: params[:template])
-      @scenario = @template.to_scenario(@scenario) #to_param)
+      @scenario = @template.to_scenario(@scenario)
     end
   end
 
@@ -24,8 +37,10 @@ class DraftsController < ApplicationController
 
       if params[:draft]
         create_action = Proc.new { |x| Scenario.create_draft(x) }
+        p 'create_draft'
       elsif params[:publish]
         create_action = Proc.new { |x| Scenario.create_live(x) }
+        p 'create_live'
       else
         raise "Illegal commit value of " + params[:commit]
       end
@@ -54,12 +69,6 @@ class DraftsController < ApplicationController
     #   flash that list
   end
 
-  def edit
-    new(Scenario.find_by_hashid(params[:id]))
-    render :new
-    # need to modify the new page to indicate it is editing
-  end
-
   def show
     redirect_to edit_draft_path
   end
@@ -70,26 +79,41 @@ class DraftsController < ApplicationController
       if params[:draft]
         @scenario.update_draft(scenario_params)
       elsif params[:publish]
-        @scenario.update_live(scenario_params.merge({
-          published_at: Time.new
-        }))
+        @scenario.update_live(scenario_params)
+      elsif params[:update]
+        @scenario.update(scenario_params)
       else
         raise "Illegal commit value of " + params[:commit]
       end
 
-      @scenario.scenario_steps.each do |step|
-        step.destroy
+      if params[:update]
+        # only update the text of them in order
+        scenario_steps_params.each do |new_step|
+          old_step = @scenario.scenario_steps.find_by(step_order: new_step[:step_order])
+          old_step.description = new_step[:description]
+          old_step.url = new_step[:url]
+          old_step.save()
+        end
+
+      else
+        @scenario.scenario_steps.each do |step|
+          step.destroy
+        end
+
+        @scenario.scenario_steps.create(scenario_steps_params)
       end
 
-      @scenario.scenario_steps.create(scenario_steps_params)
     end
 
     if params[:publish]
       flash[:info] = "<strong>Study Published</strong> - Your study link is now ready to be shared with your users."
       redirect_to result_path(@scenario)
-    else
+    elsif params[:draft]
       flash[:info] = "<strong>Draft Saved</strong> - #{@scenario.title}"
       redirect_to drafts_path
+    elsif params[:update]
+      flash[:info] = "<strong>Study Updated</strong> - Your study link is now ready to be shared with your users."
+      redirect_to result_path(@scenario)
     end
 
     # need to update in place
